Move fromLiteral functions back into their classes
